requirements.txt
requests
streamlit
gspread
oauth2client
python-dotenv